extract inner class noise
