+++ system_test/src/test/java/com/spredfast/kafka/connect/s3/S3ConnectorIntegrationTest.java
@@ -24,17 +24,9 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import com.google.common.io.Files;
-import com.netflix.curator.test.InstanceSpec;
 import com.spotify.docker.client.DefaultDockerClient;
-import com.spotify.docker.client.DockerClient;
 import com.spotify.docker.client.exceptions.DockerCertificateException;
-import com.spotify.docker.client.exceptions.DockerException;
-import com.spotify.docker.client.messages.ContainerConfig;
-import com.spotify.docker.client.messages.ContainerCreation;
-import com.spotify.docker.client.messages.HostConfig;
-import com.spotify.docker.client.messages.PortBinding;
 import com.spredfast.kafka.connect.s3.sink.S3SinkConnector;
 import com.spredfast.kafka.test.KafkaIntegrationTests;
 
@@ -140,47 +132,6 @@ public class S3ConnectorIntegrationTest {
 			.put("s3.endpoint", s3.getEndpoint())
 			.put("s3.path_style", "true") // necessary for FakeS3
 			.build();
-	}
-
-	public static class FakeS3 {
-		private static final String IMAGE = "lphoward/fake-s3";
-		private static final String CONTAINER_PORT = "4569";
-		private final int hostPort;
-		private final ContainerCreation container;
-
-
-		private FakeS3(int hostPort, ContainerCreation container) {
-			this.hostPort = hostPort;
-			this.container = container;
-		}
-
-		public static FakeS3 create(DockerClient dockerClient) throws DockerException, InterruptedException {
-			// make sure we have the image
-			dockerClient.pull(IMAGE, System.err::println);
-
-			// bind a fakes3 image to a random host port
-			int port = InstanceSpec.getRandomPort();
-			return new FakeS3(port, dockerClient.createContainer(ContainerConfig.builder()
-				.hostConfig(HostConfig.builder().portBindings(ImmutableMap.of(
-					CONTAINER_PORT, ImmutableList.of(PortBinding.of("0.0.0.0", port))
-				)).build())
-				.image(IMAGE)
-				.exposedPorts(ImmutableSet.of(CONTAINER_PORT))
-				.build()));
-		}
-
-		public void start(DockerClient dockerClient) throws DockerException, InterruptedException {
-			dockerClient.startContainer(container.id());
-		}
-
-		public void close(DockerClient dockerClient) throws DockerException, InterruptedException {
-			dockerClient.stopContainer(container.id(), 10);
-			dockerClient.removeContainer(container.id());
-		}
-
-		public String getEndpoint() {
-			return "http://localhost:" + hostPort;
-		}
 	}
 
 	private void givenRecords(String originalTopic, Producer<String, String> producer) throws InterruptedException, java.util.concurrent.ExecutionException, java.util.concurrent.TimeoutException {
